requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
finance-datareader==0.9.102
fonttools==4.60.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.9.4
multitasking==0.0.12
narwhals==2.5.0
numpy==2.0.2
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
requests-file==3.0.1
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.66
zipp==3.23.0
